predicthub_backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-environ==0.11.2
psycopg2-binary==2.9.10
drf-yasg==1.21.7
strawberry-graphql-django==0.15.0
celery==5.4.0
redis==5.2.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

